% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-save.R
\name{saveDelayedObject,array-method}
\alias{saveDelayedObject,array-method}
\alias{saveDelayedObject,DelayedArray-method}
\alias{saveDelayedObject,CsparseMatrix-method}
\title{Saving simple seed classes}
\usage{
\S4method{saveDelayedObject}{array}(x, file, name)

\S4method{saveDelayedObject}{DelayedArray}(x, file, name)

\S4method{saveDelayedObject}{CsparseMatrix}(x, file, name)
}
\arguments{
\item{x}{An R object of the indicated class.}

\item{file}{String containing the path to a HDF5 file.}

\item{name}{String containing the name of the group to save into.}
}
\value{
A \code{NULL}, invisibly.
A group is created at \code{name} containing the contents of \code{x}.
}
\description{
Methods to save simple seed classes into the delayed operation file.
See the \dQuote{Specification} vignette for details on the layout.
}
\examples{
library(HDF5Array)
X <- rsparsematrix(100, 20, 0.1)
Y <- DelayedArray(X)
temp <- tempfile(fileext=".h5")
saveDelayed(Y, temp)
rhdf5::h5ls(temp)
loadDelayed(temp)

}
\author{
Aaron Lun
}
