% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chihaya-utils}
\alias{chihaya-utils}
\alias{.saveList}
\alias{.loadList}
\alias{.labelArrayGroup}
\alias{.labelOperationGroup}
\alias{.writeScalar}
\title{Developer utilities for custom extensions}
\arguments{
\item{file}{String containing a path to a file.}

\item{name}{String containing the name of the object inside the file.
This should be a full path from the root of the file, unless \code{parent} is provided, in which case it may be the name of the child.}

\item{x}{The object to save.
\itemize{ 
\item For \code{.pickArrayType}, this should be an array-like object.
\item For \code{.saveList}, this should be a list.
\item For \code{.saveScalar}, this should be a length-1 integer, logical, character or double vector.
}}

\item{parent}{String containing the name of the parent containing the child \code{name}.}

\item{vectors.only}{Logical scalar indicating whether elements of \code{x} should be saved and loaded as 1-d arrays rather than seeds.}

\item{op}{String containing the name of the delayed operation to use to label the group.}

\item{arr}{String containing the name of the delayed array to use to label the group.}
}
\value{
\code{.saveList} and \code{.saveScalar} will write \code{x} to file, returning \code{NULL} invisibly.

\code{.labelArrayGroup} and \code{.labelOperationGroup} will apply the label to the specified group, returning \code{NULL} invisibly.

\code{.loadList} will return a list containing the contents of \code{name}.
This is guaranteed to contain only vectors (or fail) if \code{vectors.only=TRUE}.

\code{.pickArrayType} will return a string containing the \pkg{chihaya} type for an array-like \code{x}.
}
\description{
Convenience utilities for extending the \pkg{chihaya} format with \dQuote{custom} seeds or operations.
These should only be used by package developers.
}
\author{
Aaron Lun
}
