% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDF5ArraySeed.R
\name{saveLayer,HDF5ArraySeed-method}
\alias{saveLayer,HDF5ArraySeed-method}
\alias{saveLayer,H5SparseMatrixSeed-method}
\alias{saveLayer,TENxMatrixSeed-method}
\alias{saveLayer,Dense_H5ADMatrixSeed-method}
\alias{saveLayer,CSR_H5ADMatrixSeed-method}
\alias{saveLayer,CSC_H5ADMatrixSeed-method}
\title{Save HDF5-based seeds}
\usage{
\S4method{saveLayer}{HDF5ArraySeed}(x, file, name)

\S4method{saveLayer}{H5SparseMatrixSeed}(x, file, name)

\S4method{saveLayer}{TENxMatrixSeed}(x, file, name)

\S4method{saveLayer}{Dense_H5ADMatrixSeed}(x, file, name)

\S4method{saveLayer}{CSR_H5ADMatrixSeed}(x, file, name)

\S4method{saveLayer}{CSC_H5ADMatrixSeed}(x, file, name)
}
\arguments{
\item{x}{A \linkS4class{HDF5ArraySeed} or \linkS4class{H5SparseMatrix} object or subclass thereof.}

\item{file}{String containing the path to a HDF5 file.}

\item{name}{String containing the name of the group to save into.}
}
\value{
A \code{NULL}, invisibly.
A group is created at \code{name} containing the contents of the HDF5-based seed.
}
\description{
Save \linkS4class{HDF5ArraySeed} or \linkS4class{H5SparseMatrix} objects or their subclasses.
See the \dQuote{Specification} vignette for details on the layout.
}
\examples{
library(HDF5Array)
X <- writeHDF5Array(matrix(runif(100), ncol=20))
Y <- X + 1
temp <- tempfile(fileext=".h5")
saveDelayed(Y, temp)
rhdf5::h5ls(temp)
loadDelayed(temp)

}
\author{
Aaron Lun
}
