% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/simple.R
\name{saveDelayedObject}
\alias{saveDelayedObject}
\alias{saveDelayedObject,DelayedArray-method}
\title{Save a delayed object}
\usage{
saveDelayedObject(x, file, name)

\S4method{saveDelayedObject}{DelayedArray}(x, file, name)
}
\arguments{
\item{x}{An R object containing a delayed operation or seed class.}

\item{file}{String containing the path to a HDF5 file.}

\item{name}{String containing the name of the group to save into.}
}
\value{
A \code{NULL} is returned invisibly.
A group is created at \code{name} inside \code{file} and the delayed operation is saved within.
}
\description{
Saves a delayed object recursively.
}
\details{
The \code{\link{saveDelayedObject}} generic is intended for developers to create methods for new operations.
End-users should use the \code{\link{saveDelayed}} function instead.

The \linkS4class{DelayedArray} method will simply extract the seed and use it to call \code{\link{saveDelayedObject}} again.
}
\author{
Aaron Lun
}
