% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedAbind.R
\name{saveLayer,DelayedAbind-method}
\alias{saveLayer,DelayedAbind-method}
\title{Saving a DelayedAbind}
\usage{
\S4method{saveLayer}{DelayedAbind}(x, file, name)
}
\arguments{
\item{x}{A \linkS4class{DelayedAbind} object.}

\item{file}{String containing the path to a HDF5 file.}

\item{name}{String containing the name of the group to save into.}
}
\value{
A \code{NULL}, invisibly.
A group is created at \code{name} containing the contents of the DelayedUnaryIsoOpStack.
}
\description{
Save a \linkS4class{DelayedAbind} object.
}
\details{
The DelayedAbind class is represented on file by several elements within the \code{name}d group.
\itemize{
\item \code{type}, a character dataset with two elements.
The first is set to \code{"OPERATION"} and the second is set to \code{"COMBINE"}.
\item \code{along}, a dataset specifying the dimension to combine along.
This should be integer and of length 1.
\item \code{seeds}, a group describing the seeds to combine.
This contains additional groups or datasets, named by the order in which they are to be combined.
}
}
\examples{
X <- DelayedArray(matrix(runif(100), ncol=20))
Y <- cbind(X, X)
temp <- tempfile(fileext=".h5")
saveDelayedOps(Y, temp)
rhdf5::h5ls(temp)

}
\author{
Aaron Lun
}
