% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedSetDimnames.R
\name{saveLayer,DelayedSetDimnames-method}
\alias{saveLayer,DelayedSetDimnames-method}
\title{Saving a DelayedSetDimnames}
\usage{
\S4method{saveLayer}{DelayedSetDimnames}(x, file, name)
}
\arguments{
\item{x}{A \linkS4class{DelayedSetDimnames} object.}

\item{file}{String containing the path to a HDF5 file.}

\item{name}{String containing the name of the group to save into.}
}
\value{
A \code{NULL}, invisibly.
A group is created at \code{name} containing the contents of the DelayedSetDimnames.
}
\description{
Save a \linkS4class{DelayedSetDimnames} object.
}
\details{
The DelayedSetDimnames class is represented on file by several elements within the \code{name}d group.
\itemize{
\item \code{type}, a character dataset with two elements.
The first is set to \code{"OPERATION"} and the second is set to \code{"DIMNAMES"}.
\item \code{dimnames}, a group containing \code{number}, the number of dimensions,
and any number of string datasets with names in [1, number].
Each dataset contains names along the named dimension of the array.
\item \code{seed}, a group or dataset containing the seed to be named.
}
}
\examples{
X <- DelayedArray(matrix(runif(100), ncol=20))
Y <- cbind(X, X)
temp <- tempfile(fileext=".h5")
saveDelayedOps(Y, temp)
rhdf5::h5ls(temp)

}
\author{
Aaron Lun
}
