% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedSubassign.R
\name{saveLayer,DelayedSubassign-method}
\alias{saveLayer,DelayedSubassign-method}
\title{Saving a DelayedSubassign}
\usage{
\S4method{saveLayer}{DelayedSubassign}(x, file, name)
}
\arguments{
\item{x}{A \linkS4class{DelayedSubassign} object.}

\item{file}{String containing the path to a HDF5 file.}

\item{name}{String containing the name of the group to save into.}
}
\value{
A \code{NULL}, invisibly.
A group is created at \code{name} containing the contents of the DelayedSubassign.
}
\description{
Save a \linkS4class{DelayedSubassign} object into a HDF5 file.
}
\details{
The DelayedSubassign class is represented on file by several elements within the \code{name}d group.
\itemize{
\item \code{type}, a character dataset with two elements.
The first is set to \code{"OPERATION"} and the second is set to \code{"SUBSET"}.
\item \code{index}, a group containing \code{number}, the number of dimensions,
and any number of integer datasets with names in [1, number].
Each dataset contains 1-based indices to subset the named dimension of the array.
\item \code{value}, a group or dataset containing the value to use for replacement.
This may potentially be another seed-like structure.
\item \code{seed}, a group or dataset containing the seed to assign the replacement value to.
}
}
\examples{
X <- DelayedArray(matrix(runif(100), ncol=20))
X[1:2,3:5] <- matrix(-runif(6), ncol=3)
temp <- tempfile(fileext=".h5")
saveDelayedOps(Y, temp)
rhdf5::h5ls(temp)

}
\author{
Aaron Lun
}
