% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedUnaryIsoOpStack.R
\name{saveLayer,DelayedUnaryIsoOpStack-method}
\alias{saveLayer,DelayedUnaryIsoOpStack-method}
\title{Saving a DelayedUnaryIsoOpStack}
\usage{
\S4method{saveLayer}{DelayedUnaryIsoOpStack}(x, file, name)
}
\arguments{
\item{x}{A \linkS4class{DelayedUnaryIsoOpStack} object.}

\item{file}{String containing the path to a HDF5 file.}

\item{name}{String containing the name of the group to save into.}
}
\value{
A \code{NULL}, invisibly.
A group is created at \code{name} containing the contents of the DelayedUnaryIsoOpStack.
}
\description{
Save a \linkS4class{DelayedUnaryIsoOpStack} object into a HDF5 file.
}
\details{
The DelayedUnaryIsoOpStack class is represented on file by several elements within the \code{name}d group.
\itemize{
\item \code{type}, a character dataset with two elements.
The first is set to \code{"OPERATION"} and the second is set to \code{"UNARY"}.
\item \code{operations}, a group containing information about the operations.
Each operation is represented by a group that is named by the execution order, i.e., \code{"1"} is performed first, then \code{"2"}, and so on.
\item \code{seed}, a group or dataset containing the seed to be operated on.
}

Each group in \code{operations} contains:
\itemize{
\item \code{name}, a character dataset with a single element naming the unary operation to be performed.
The seed is assumed to be the first argument in this operation.
Note that the unary-ness refers to the fact that we only need to consider one seed; there may still be multiple arguments.
\item (optional) \code{arguments}, a group containing additional arguments to use in the operation, in addition to the \code{seed}.
Each argument is a dataset inside the group where the name of the dataset is the name of the argument.
If this group is not present, no further arguments are necessary.
}

Valid operation names in \code{name} are those listed for \code{\link{Ops}}, \code{\link{Math}} and \code{\link{Math2}}.
The format of \code{arguments} depends on the nature of \code{name}.
\itemize{
\item For operations listed in \code{\link{Ops}}, \code{arguments} will contain a \code{side} dataset of length 1.
This will either say \code{"left"} or \code{"right"}, indicating whether the operation is applied to the left or the right side of the seed.
\code{arguments} will also contain \code{value} containing the value of the non-seed operand.
\item For operations listed in \code{\link{Math2}}, \code{arguments} will contain a \code{digits} dataset.
This contains the digits to be used in the various rounding operations.
\item For operations listed in \code{\link{Math}}, \code{arguments} is ignored and should contain no arguments.
}
}
\examples{
X <- DelayedArray(matrix(runif(100), ncol=20))
Y <- log2(X + 10)
temp <- tempfile(fileext=".h5")
saveDelayedOps(Y, temp)
rhdf5::h5ls(temp)

}
\author{
Aaron Lun
}
