% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{saveLayer}
\alias{saveLayer}
\title{Save a delayed layer}
\usage{
saveLayer(x, file, name)
}
\arguments{
\item{x}{An R object containing a delayed operation or seed class.}

\item{file}{String containing the path to a HDF5 file.}

\item{name}{String containing the name of the group to save into.}
}
\value{
A \code{NULL} is returned invisibly.
A group is created at \code{name} inside \code{file} and the delayed operation is saved within.
}
\description{
Saves a delayed layer, either an operation or a seed array.
}
\details{
The \code{\link{saveLayer}} generic is intended for developers to create methods for new operations.
End-users should use the \code{\link{saveDelayed}} function instead.
}
\author{
Aaron Lun
}
