% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{utils}
\alias{utils}
\alias{.saveList}
\alias{.loadList}
\alias{.labelSeedGroup}
\alias{.labelOperationGroup}
\alias{.writeScalar}
\title{Developer utilities for custom extensions}
\description{
Convenience utilities for extending the \pkg{chihaya} format with \dQuote{custom} seeds or operations.
These should only be used by package developers.
}
