% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate an artifact}
\usage{
validate(path, name)
}
\arguments{
\item{path}{String containing the path to the HDF5 file.}

\item{name}{String containing the name of the delayed object inside the file.}
}
\value{
\code{NULL} if there are no problems, otherwise an error is raised.
}
\description{
Validate the delayed objects inside a HDF5 file.
This is automatically run at the end of every \code{\link{saveDelayed}} call to check the integrity of the saved files.
See \url{https://ltla.github.io/chihaya} for more details.
}
\seealso{
See \url{https://ltla.github.io/chihaya} for the specification.
}
\author{
Aaron Lun
}
