% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-save.R
\name{saveDelayedObject,array-method}
\alias{saveDelayedObject,array-method}
\alias{saveDelayedObject,DelayedArray-method}
\alias{saveDelayedObject,CsparseMatrix-method}
\alias{saveDelayedObject,ANY-method}
\title{Saving simple seed classes}
\usage{
\S4method{saveDelayedObject}{array}(x, file, name)

\S4method{saveDelayedObject}{DelayedArray}(x, file, name)

\S4method{saveDelayedObject}{CsparseMatrix}(x, file, name)

\S4method{saveDelayedObject}{ANY}(x, file, name)
}
\arguments{
\item{x}{An R object of the indicated class.}

\item{file}{String containing the path to a HDF5 file.}

\item{name}{String containing the name of the group to save into.}
}
\value{
A \code{NULL}, invisibly.
A group is created at \code{name} containing the contents of \code{x}.
}
\description{
Methods to save simple seed classes into the delayed operation file.
See \dQuote{Dense arrays} and \dQuote{Sparse matrices} at \url{https://ltla.github.io/chihaya} for more details.
}
\details{
The ANY method will dispatch to classes that are implemented in other packages:
\itemize{
\item If \code{x} is a LowRankMatrixSeed from the \pkg{BiocSingular} package, it is handled as a delayed matrix product.
}
}
\examples{
# Saving a sparse matrix.
X <- rsparsematrix(100, 20, 0.1)
Y <- DelayedArray(X)
temp <- tempfile(fileext=".h5")
saveDelayed(Y, temp)
rhdf5::h5ls(temp)
loadDelayed(temp)

# Saving a matrix product.
library(BiocSingular)
left <- matrix(rnorm(100000), ncol=20)
right <- matrix(rnorm(50000), ncol=20)
thing <- LowRankMatrix(left, right)
temp <- tempfile()
saveDelayed(thing, temp)
rhdf5::h5ls(temp)
loadDelayed(temp)

}
\author{
Aaron Lun
}
