% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedSubset.R
\name{saveLayer,DelayedSubset-method}
\alias{saveLayer,DelayedSubset-method}
\title{Saving a DelayedSubset}
\usage{
\S4method{saveLayer}{DelayedSubset}(x, file, name)
}
\arguments{
\item{x}{A \linkS4class{DelayedSubset} object.}

\item{file}{String containing the path to a HDF5 file.}

\item{name}{String containing the name of the group to save into.}
}
\value{
A \code{NULL}, invisibly.
A group is created at \code{name} containing the contents of the DelayedSubset.
}
\description{
Save a \linkS4class{DelayedSubset} object into a HDF5 file.
}
\details{
The DelayedSubset class is represented on file by several elements within the \code{name}d group.
\itemize{
\item \code{type}, a character dataset with two elements.
The first is set to \code{"OPERATION"} and the second is set to \code{"SUBSET"}.
\item \code{index}, a group containing \code{number}, the number of dimensions,
and any number of integer datasets with names in [1, number].
Each dataset contains 1-based indices to subset the named dimension of the array.
\item \code{seed}, a group or dataset containing the seed to be subsetted.
}
}
\examples{
X <- DelayedArray(matrix(runif(100), ncol=20))
Y <- X[1:2,3:5]
temp <- tempfile(fileext=".h5")
saveDelayedOps(Y, temp)
rhdf5::h5ls(temp)

}
\author{
Aaron Lun
}
